Simplify `CooperOptimizerState`
