+++ cooper/optim/optimizer_state.py
@@ -4,10 +4,6 @@ import dataclasses
 from collections.abc import Sequence
 from typing import Dict, Optional
 
-from cooper.utils import validate_state_dicts
-
-from .types import AlternationType
-
 
 @dataclasses.dataclass
 class CooperOptimizerState:
@@ -19,35 +15,7 @@ class CooperOptimizerState:
     Args:
         primal_optimizer_states: State dicts for the primal optimizers.
         dual_optimizer_states: State dicts for the dual optimizers.
-        multiplier_states: State dicts for the multipliers.
-        extrapolation: Flag indicating if the optimizer performs extrapolation updates.
-        alternation_type: Choice of alternation strategy.
     """
 
     primal_optimizer_states: Sequence[Dict]
     dual_optimizer_states: Optional[Sequence[Dict]] = None
-    multiplier_states: Optional[Sequence[Dict]] = None
-    extrapolation: bool = False
-    alternation_type: AlternationType = AlternationType.FALSE
-
-    def asdict(self):
-        return dataclasses.asdict(self)
-
-    def __eq__(self, other):
-        assert isinstance(other, CooperOptimizerState)
-
-        flag_names = ["extrapolation", "alternation_type"]
-        for flag_name in flag_names:
-            if getattr(self, flag_name) != getattr(other, flag_name):
-                return False
-
-        state_dict_names = ["primal_optimizer_states", "dual_optimizer_states", "multiplier_states"]
-        for state_dict_name in state_dict_names:
-            try:
-                dicts_match = validate_state_dicts(getattr(self, state_dict_name), getattr(other, state_dict_name))
-            except:  # noqa: E722
-                dicts_match = False
-            if not dicts_match:
-                return False
-
-        return True
