+++ cooper/utils/__init__.py
@@ -1,2 +1,2 @@
 from .typing import OneOrSequence
-from .utils import ensure_sequence, validate_state_dicts
+from .utils import ensure_sequence
